refactor <LottoParser>: 기능 분리
Validator 생성해서 파싱과 검증 로직 분리
